change choose winner rule

+++ contracts/lottery/src/error.rs
@@ -52,7 +52,7 @@ pub enum ContractError {
     PaymentNotEnough { amount: Uint128 },
 
     #[error("The lottery player is full: {max_players}")]
-    PlayerExceededMaximum { max_players: u32 },
+    PlayerExceededMaximum { max_players: u64 },
 
     #[error("The lottery is activing, can't draw")]
     LotteryIsActiving {},
